problem to connect module to app module and to index.tsx
