+++ project/src/components/app/app.tsx
@@ -1,5 +1,10 @@
-function App(): JSX.Element {
-  return <p>Hello, world!</p>;
+import { MovieProps } from '../const/const';
+import PageContent from '../main/main-page-content';
+
+function App(film: MovieProps): JSX.Element {
+  return (
+    <PageContent film = {film} />
+  );
 }
 
 export default App;
