+++ project/src/components/main/main-page-content.tsx
@@ -1,17 +1,19 @@
-import { filmTs } from '../const/const';
+import { MovieProps } from '../const/const';
 import MainGenreFilters from './main-genre-filters';
 import MovieCatalog from './main-movie-catalog';
 import Logo from '../general/logo';
 const CATALOG = 'Catalog';
 
-function PageContent(film: filmTs): JSX.Element {
+function PageContent(film: MovieProps): JSX.Element {
   return(
     <div className="page-content">
       <section className="catalog">
         <h2 className="catalog__title visually-hidden">{CATALOG}</h2>
         {MainGenreFilters(film)}
         <div className="catalog__films-list">
-          {MovieCatalog(film)}
+          {new Array(20)
+            .fill('')
+            .map((line, index) => MovieCatalog(film, index))}
         </div>
         <div className="catalog__more">
           <button className="catalog__button" type="button">Show more</button>
